Relocate drake_gym into bindings. Rewrite READMEs to reflect movement (and additional edits for clarity)
