+++ bindings/pydrake/examples/gym/BUILD.bazel
@@ -31,13 +31,12 @@ py_library(
     name = "box_flipup",
     srcs = ["envs/box_flipup/box_flipup.py"],
     data = [
-        "//drake_gym/models",
+        "//bindings/pydrake/examples/gym/models",
     ],
     visibility = ["//visibility:public"],
     deps = [
         ":scenario_helpers",
         "//bindings/pydrake",
-        "//drake_gym",
     ],
 )
 
@@ -45,14 +44,13 @@ py_library(
     name = "cart_pole",
     srcs = ["envs/cart_pole/cart_pole.py"],
     data = [
-        "//drake_gym/models",
+        "//bindings/pydrake/examples/gym/models",
     ],
     visibility = ["//visibility:public"],
     deps = [
         ":named_view_helpers",
         ":scenario_helpers",
         "//bindings/pydrake",
-        "//drake_gym",
     ],
 )
 
@@ -60,12 +58,12 @@ py_binary(
     name = "train_box_flipup",
     srcs = ["train_box_flipup.py"],
     data = [
-        "//drake_gym/models",
+        "//bindings/pydrake/examples/gym/models",
     ],
     visibility = ["//visibility:public"],
     deps = [
-        "//drake_gym",
-        "//drake_gym/examples:box_flipup",
+        "//bindings/pydrake",
+        "@gym_py",
         "@stable_baselines3_internal//:stable_baselines3",
     ],
 )
@@ -74,12 +72,12 @@ py_binary(
     name = "train_cart_pole",
     srcs = ["train_cart_pole.py"],
     data = [
-        "//drake_gym/models",
+        "//bindings/pydrake/examples/gym/models",
     ],
     visibility = ["//visibility:public"],
     deps = [
-        "//drake_gym",
-        "//drake_gym/examples:cart_pole",
+        "//bindings/pydrake",
+        "@gym_py",
         "@stable_baselines3_internal//:stable_baselines3",
     ],
 )
@@ -88,12 +86,11 @@ py_binary(
     name = "play_cart_pole",
     srcs = ["play_cart_pole.py"],
     data = [
-        "//drake_gym/models",
+        "//bindings/pydrake/examples/gym/models",
     ],
     visibility = ["//visibility:public"],
     deps = [
-        "//drake_gym",
-        "//drake_gym/examples:cart_pole",
+        "//bindings/pydrake/examples/gym:cart_pole",
     ],
 )
 
@@ -105,8 +102,8 @@ py_test(
     args = ["--test"],
     main = "train_box_flipup.py",
     deps = [
-        "//drake_gym",
-        "//drake_gym/examples:box_flipup",
+        "//bindings/pydrake/examples/gym:box_flipup",
+        "//bindings/pydrake/gym",
         "@stable_baselines3_internal//:stable_baselines3",
     ],
 )
@@ -117,8 +114,8 @@ py_test(
     args = ["--test"],
     main = "train_cart_pole.py",
     deps = [
-        "//drake_gym",
-        "//drake_gym/examples:cart_pole",
+        "//bindings/pydrake/examples/gym:cart_pole",
+        "//bindings/pydrake/gym",
         "@stable_baselines3_internal//:stable_baselines3",
     ],
 )
