+++ bindings/pydrake/examples/gym/README.md
@@ -1,0 +1,8 @@
+Examples for Drake Gym
+======================
+
+This directory contains some simple examples using drake_gym.  For more detail
+see the documentation for [drake_gym itself](/bindings/pydrake/gym/README.md).
+
+THIS FEATURE IS EXPERIMENTAL.  Development is ongoing and no guarantees
+against deprecation are provided for any file under this directory.
