+++ bindings/pydrake/examples/gym/models/cartpole_BSA.sdf
