+++ bindings/pydrake/examples/gym/models/planar_bin.sdf
