+++ bindings/pydrake/examples/gym/named_view_helpers.py
