+++ bindings/pydrake/examples/gym/scenario_helpers.py
