+++ bindings/pydrake/examples/gym/train_box_flipup.py
